requirements.txt
absl-py==2.0.0
aiobotocore==2.9.0
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.42.0
blessed==1.20.0
boto3==1.33.13
botocore==1.33.13
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
croniter==1.4.1
cycler==0.12.1
dateutils==0.6.12
deepdiff==6.7.1
docker==6.1.3
docstring-parser==0.15
editor==1.6.5
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.26.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.14.0
huggingface-hub==0.20.2
hydra-core==1.3.2
idna==3.6
importlib-resources==6.1.1
inquirer==3.2.1
Jinja2==3.1.3
jmespath==1.0.1
jsonargparse==4.27.1
kiwisolver==1.4.5
lightning==2.1.3
lightning-api-access==0.0.5
lightning-cloud==0.5.57
lightning-fabric==2.1.3
lightning-utilities==0.10.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
ordered-set==4.1.0
packaging==23.2
Pillow==10.1.0
pip==23.3.1
protobuf==4.23.4
psutil==5.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
runs==1.2.0
s3fs==2023.12.2
s3transfer==0.8.2
safetensors==0.4.1
scipy==1.11.4
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
timm==0.9.12
tokenizers==0.15.0
torch==2.1.2
torchmetrics==1.3.0
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
triton==2.1.0
types-python-dateutil==2.8.19.20240106
typeshed-client==2.4.0
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.25.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
xmod==1.8.1
yarl==1.9.4
